requirements.txt
psycopg2-binary
sqlalchemy
python-dotenv
docling
tiktoken
requests
google-generativeai
pandas
fastapi
uvicorn
python-multipart
